use new job billing table in init
