+++ apps/mis-web/src/pageComponents/init/InitJobBillingTable.tsx
@@ -15,21 +15,31 @@ import { useCallback } from "react";
 import { useAsync } from "react-async";
 import { api } from "src/apis";
 import { Centered } from "src/components/layouts";
-import { EditableJobBillingTable } from "src/pageComponents/job/EditableJobBillingTable";
+import { ManageJobBillingTable } from "src/pageComponents/job/ManageJobBillingTable";
+import { getActiveBillingItems } from "src/utils/job";
 
-export const EditJobPriceTableForm: React.FC = () => {
-
+export const InitJobBillingTable: React.FC = () => {
   const { data, isLoading, reload } = useAsync({ promiseFn: useCallback(async () => {
-    return await api.getBillingTable({ query: { } }).then((x) => x.items);
+    const sourceData = await api.getBillingItems({ 
+      query: { tenant: undefined, activeOnly: false },
+    }).then((x) => x.items);
+    return getActiveBillingItems(sourceData, undefined);
   }, []) });
 
   return (
     <Centered>
       <div>
-        <Typography.Paragraph>您可以在这里设置默认作业价格表。未设置的将会以0元计费。</Typography.Paragraph>
-        <EditableJobBillingTable reload={reload} data={data} loading={isLoading} />
+        <Typography.Paragraph>
+          您可以在这里设置默认作业价格表。未设置的将会以0元计费。
+          <a onClick={reload}>刷新</a>
+        </Typography.Paragraph>
+        <ManageJobBillingTable
+          data={data}
+          loading={isLoading}
+          tenant={undefined}
+          reload={reload}
+        />
       </div>
     </Centered>
   );
-
 };
